backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-multipart==0.0.6
# openai-whisper==20231117
# torch==2.1.0
# torchvision==0.16.0
# torchaudio==2.1.0
# yt-dlp
# pytubefix
# beautifulsoup4==4.12.2
# requests==2.31.0
# moviepy==1.0.3
# langchain==0.0.340
# ollama==0.1.7
# urllib3==2.0.7
# certifi==2023.11.17
# google-generativeai==0.3.2
# openai==1.12.0
# anthropic==0.18.1
# python-dotenv==1.0.0
# Backend Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# File Uploads
python-multipart==0.0.6

# Whisper (for speech-to-text)
openai-whisper==20231117
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# YouTube Downloaders
yt-dlp
pytubefix

# Web Scraping & Requests
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==2.0.7
certifi==2023.11.17

# Video Processing
moviepy==1.0.3

# LLM Frameworks & Providers
langchain==0.0.340
ollama==0.1.7
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Environment Variable Loader
python-dotenv==1.0.0
